changed field names
Renamed TMDL_active column to TMDL_status. The status values are no longer a TRUE/FALSE data type. This was changed to align with the data type and domain values in ATTAINS, and allow inclusion of a draft in development TMDL status.
Renamed TMDL_active_note to TMDL_status_comment.
Added revision_action_id to link a non-active TMDL with it's revision.